requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
Django==3.2.5
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
twilio==6.61.0
typing-extensions==3.10.0.0
urllib3==1.26.6
